Rename ConvertFrom-Markdown
PowerShellCore 6 added a `ConvertFrom-Markdown` command, and this module's
command therefore causes a conflict.

Renaming to `ConvertFrom-GitHubMarkdown` in an effort to best identify this
as being a "GitHub" method that converts Markdown, although it unfortunately
does imply that it might only support "GitHub Flavored Markdown", where in fact
it supports both that as well as standard.

Resolves #99 - Rename ConvertFrom-Markdown