Reducing warning messages during unit tests configuration (#132)
During UT's, `Reset-GitHubConfiguration` and `Clear-GitHubAuthentication` are called frequently.
This change moves their warning messages to be in Verbose output instead, in order to reduce
the noise that is commonly seen when they're called.

Additionally, the warning that is seen at the start of running a UT test suite is now
completely suppressed if the user has modified `Tests/Config/Settings.ps1` (presumably to
have it reference accounts that they have access for). This is achieved by storing the hash of the Settings file in the code, along with documentation to remind users to update it if the Settings
file is ever permanently changed.

Finally, an additional test has been added to the CI pipeline to validate that the Settings file
content matches the hash that is present in the code.