Add automatic daily update checking (#185)
The module will now check daily with PowerShellGallery to see if there
is a newer version of the module available.  This check can be disabled
with the new `DisableUpdateCheck` config property.

The check happens asynchronously.  The web request is kicked off when
the module is loaded, but the result is only checked when the user
runs a command.

The result is logged at the Verbose level in any failure case as well
as when the module is up-to-date.  When there is a newer version
available, it will write the message out as a Warning.  The web request
will only run once per day, and the user message will only be written
once per day as well.