Added API versioning support (#379)
GitHub is introducing versioned API's.  In order to get the expected behavior of an API, you need to invoke the API with its version in the request header.  This change updates the core functions (`Invoke-GHRestMethod` and `Invoke-GHRestMethodMultipleResult`) to accept an ApiVersion as part of their param sets so that as the module continues to add new functionality, methods can request the appropriate API Version for their individual calls.

Fixes #377 

Reference: [To infinity and beyond: enabling the future of GitHub’s REST API with API versioning](https://github.blog/2022-11-28-to-infinity-and-beyond-enabling-the-future-of-githubs-rest-api-with-api-versioning/)
[API Versions](https://docs.github.com/en/rest/overview/api-versions?apiVersion=2022-11-28)