Replace `BinaryFormatter` with `PSSerializer` (#415)
Replace `BinaryFormatter` with `PSSerializer`

Since the former has been removed from PowerShell 7.4 (because it was removed from .NET 7) and the latter can accomplish the same thing.

The default depth is 1, so we need to specify. Best I could find was that the prior implementation defaulted to a depth of 64, so we're going with that.